Update G_abundance_est.py
I updated the function to be more generic. I was going to make the input an SRR folder, but that makes it more difficult to test on my own laptop because I do not have everything organized in the correct folder structure, so I think it is better to leave it generic, and have the more specific code in the main file. I will be writing code for this in my tester.